cleanup: incorporate suggestions from @kennyweiss
